Refactor some activity extensions so that they can be used outside the app module
